pyproject.toml
[tool.poetry]
name = "cardgame"
version = "0.1.0"
description = ""
authors = ["roycehoe <royceinbox@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
black = "^21.6b0"
mypy = "^0.910"
neovim = "^0.3.1"
pytest = "^6.2.4"
frozendict = "^2.0.3"
pytest-cov = "^2.12.1"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.25"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
pydantic = "^1.8.2"
sqlalchemy-stubs = "^0.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
requests = "^2.26.0"
pytest-freezegun = "^0.4.2"
freezegun = "^1.1.0"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
pytest-mypy = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

